Clean up taxonomy list functions and remove the_content filter
This plugin is intended for data structure and access, not presentation. It was inappropriate to filter the_content to include artwork images and data: presentational logic will be handled theme-side for now.

With the removal of the content filtering, many of the helper methods created to handle the custom taxonomies used within artgallery are no longer required. Removing them and adding a generic, theme-accessible helper for plain-text taxonomy rendering serves all the same purposes with much less fuss.